# Optimize parser a little