--- a/Cargo.toml
+++ b/Cargo.toml
@@ -8,13 +8,18 @@ edition = "2021"
 [dependencies]
 tracy-client = { version = "0.17.3", default-features = false } # for tracy v0.11.1
 
-[profile.profile]
+[profile.profiling]
 inherits = "release"
 debug = true
 
 [features]
 # By default, profiling is disabled. Enable it by the "profile" feature
-profile = ["tracy-client/enable", "tracy-client/flush-on-exit"]
+profiling = [
+    "tracy-client/enable",
+    "tracy-client/flush-on-exit",
+    "tracy-client/sampling",
+    "tracy-client/code-transfer",
+]
 
 [lib]
 name = "new_nu_parser"