--- a/src/main.rs
+++ b/src/main.rs
@@ -7,8 +7,19 @@ use new_nu_parser::typechecker::Typechecker;
 
 fn main() {
     let mut compiler = Compiler::new();
+    let mut do_print = true;
+
+    for arg in std::env::args().skip(1) {
+        if arg == "--no-print" {
+            do_print = false;
+        }
+    }
 
     for fname in std::env::args().skip(1) {
+        if fname == "--no-print" {
+            continue;
+        }
+
         let contents = std::fs::read(&fname);
 
         let Ok(contents) = contents else {
@@ -22,7 +33,9 @@ fn main() {
         let parser = Parser::new(compiler, span_offset);
         compiler = parser.parse();
 
-        compiler.print();
+        if do_print {
+            compiler.print();
+        }
 
         if !compiler.errors.is_empty() {
             exit(1);
@@ -30,7 +43,10 @@ fn main() {
 
         let mut resolver = Resolver::new(&compiler);
         resolver.resolve();
-        resolver.print();
+
+        if do_print {
+            resolver.print();
+        }
 
         compiler.merge_name_bindings(resolver.to_name_bindings());
 
@@ -40,7 +56,10 @@ fn main() {
 
         let mut typechecker = Typechecker::new(&compiler);
         typechecker.typecheck();
-        typechecker.print();
+
+        if do_print {
+            typechecker.print();
+        }
 
         compiler.merge_types(typechecker.to_types());
     }